--- a/package.json
+++ b/package.json
@@ -35,5 +35,8 @@
     "standard": "^16.0.0",
     "tap-spec": "^5.0.0",
     "tape": "^5.2.2"
+  },
+  "dependencies": {
+    "clones": "^1.2.0"
   }
 }