Hey, @mcollina I noticed that benchmarks when `tinysonic.parse()` was circular, it ran at around ~3m (5m/s) times slower. I added the `clones` module and it jumped back to 8m/s.

Before:
```
Jsonic Parse x 99,544 ops/sec ±12.33% (72 runs sampled)
TinySonic Parse x 2,139,987 ops/sec ±1.70% (91 runs sampled)
JSON Parse x 867,426 ops/sec ±10.97% (62 runs sampled)
Jsonic Stringify x 854,556 ops/sec ±11.08% (65 runs sampled)
TinySonic Stringify x 5,476,933 ops/sec ±7.80% (69 runs sampled)
JSON Stringify x 1,662,651 ops/sec ±8.16% (77 runs sampled)
```

After:
```
Jsonic Parse x 135,122 ops/sec ±13.77% (82 runs sampled)
TinySonic Parse x 2,120,028 ops/sec ±1.35% (90 runs sampled)
JSON Parse x 1,453,380 ops/sec ±3.39% (83 runs sampled)
Jsonic Stringify x 1,674,637 ops/sec ±0.73% (92 runs sampled)
TinySonic Stringify x 8,188,723 ops/sec ±0.78% (90 runs sampled)
JSON Stringify x 2,010,491 ops/sec ±1.32% (89 runs sampled)
```