@mcollina 2021-06-14
I don't think the clones module should make a difference, could you check again?